[Windows] Support NodePortLocal on Antrea Windows Agent
* Support NodePortLocal rules on by using NetNatStaticMapping on windows
* Support NPL agent on Windows platform
* Require the same Antrea NPL configuration as Linux

issue antrea-io#3826
